Fix  Template. Call bining.Init only if the GameObject template attached to it is active in the hierarchy.
GetComponentsInChildren doesn't do what you might expect in this case.
https://forum.unity.com/threads/getcomComponentsinchildren-false-returns-inactive-objects.501177/

Call GetComponentsInChildren and binding.Init() child objects only if the current game object is active in the hierarchy
If you called Binding.Init on an inactive object and it never Awake during its lifecycle (OnDestroy will not be called)
the binding will never be disconnected.